refactor: split grades.py into focused sub-modules (_scoring, _totals, _summary)
Each sub-module is under 100 lines. grades.py remains as re-export facade
for backward compatibility.
